fixed FOUC
